+++ src/index.ts
@@ -3,3 +3,5 @@ export { Service } from "./service.core"
 export { TyapiError } from "./error.core"
 
 export { Logger, LoggerMode } from "./logger/logger.service"
+
+export { ExpressServer } from "./server/express.service"